Named inference rule for `abs`.
Adds a named inference rule for `abs`. This is an example of how to add
a named inference rule for a function that simply propagates names on
its inputs to the outputs.
- Actually, this diff adds named inference rules for other functions as
well, but they are not enabled yet (due to named_guard=False for those
functions). Those will be enabled in a future change.

Test Plan:
- `python test/test_namedtensor.py` [namedtensor ci]

gh-metadata: pytorch pytorch 22151 gh/zou3519/61/head